[FIX] colors should now be working properly (even if a bit ugly), still need to fix infinite cylinders

+++ incs/data_structures/color.h
@@ -6,7 +6,7 @@
 /*   By: cviel <cviel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/23 17:59:41 by cviel             #+#    #+#             */
-/*   Updated: 2026/01/05 20:36:35 by cviel            ###   ########.fr       */
+/*   Updated: 2026/01/05 21:26:02 by cviel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ typedef struct s_color
 
 typedef struct s_sum_color
 {
-	uint16_t	r;
-	uint16_t	g;
-	uint16_t	b;
+	double	r;
+	double	g;
+	double	b;
 }	t_sum_color;
 
 #endif
